api_tester/requirements.txt
loguru==0.7.2
pandas==2.1.3
graphviz==0.20.3
aiohttp==3.9.1
pendulum==3.0.0
python-dotenv==1.0.0
numpy==1.26.4
